started moving indexing
